Refactor MyWhiskyShelfDbContext
* Refactor to ensure that the `DbContext` stays focused on entities and EF mechanics.
* `DistilleryReadService` added to handle units of work against the `DbContext`.
* Copy across test to run against the service instead of the context.
